Fixing bad performance of AssetManager due to writing guids map after every single asset modification

+++ Tests/PerformanceTests.cs
@@ -4,6 +4,7 @@ using Robot;
 using Robot.Scripts;
 using System.IO;
 using RobotRuntime.Perf;
+using RobotRuntime.Assets;
 
 namespace Tests
 {
@@ -23,43 +24,45 @@ namespace Tests
         private const string k_ScriptCPath = "Scripts\\C.mrb";
 
         [TestMethod]
-        public void AssetRefreshTime_100_ScriptAssets() // 22 ms
+        public void AssetRefreshTime_100_ScriptAssets() // 22 ms - 50 (guid)
         {
             int count = 100;
+            var name = "Refresh_Test_1";
 
             for (int i = 0; i < count; ++i)
                 CreateDummyScriptWithImporter("Scripts\\A" + i + ".mrb");
 
-            Profiler.Start("Refresh_Test");
+            Profiler.Start(name);
             AssetManager.Instance.Refresh();
-            Profiler.Stop("Refresh_Test");
+            Profiler.Stop(name);
 
             Assert.AreEqual(count, AssetManager.Instance.Assets.Count(), "Asset count missmatch");
 
-            var timeTaken = Profiler.Instace.CopyNodes()["Refresh_Test"][0].Time;
+            var timeTaken = Profiler.Instace.CopyNodes()[name][0].Time;
             System.Diagnostics.Debug.WriteLine("Asset Refresh on 100 entries took: " + timeTaken + " ms.");
 
-            Assert.IsTrue(timeTaken < 31, "This test took 40% longer than usual");
+            Assert.IsTrue(timeTaken < 80, "This test took 40% longer than usual");
         }
 
         [TestMethod]
-        public void AssetRefreshTime_1000_ScriptAssets() // 120 ms
+        public void AssetRefreshTime_1000_ScriptAssets() // 120 ms -- 180 (guid)
         {
             int count = 1000;
+            var name = "Refresh_Test_2";
 
             for (int i = 0; i < count; ++i)
                 CreateDummyScriptWithImporter("Scripts\\A" + i + ".mrb");
 
-            Profiler.Start("Refresh_Test");
+            Profiler.Start(name);
             AssetManager.Instance.Refresh();
-            Profiler.Stop("Refresh_Test");
+            Profiler.Stop(name);
 
             Assert.AreEqual(count, AssetManager.Instance.Assets.Count(), "Asset count missmatch");
 
-            var timeTaken = Profiler.Instace.CopyNodes()["Refresh_Test"][0].Time;
+            var timeTaken = Profiler.Instace.CopyNodes()[name][0].Time;
             System.Diagnostics.Debug.WriteLine("Asset Refresh on 1000 entries took: " + timeTaken + " ms.");
 
-            Assert.IsTrue(timeTaken < 165, "This test took 40% longer than usual");
+            Assert.IsTrue(timeTaken < 250, "This test took 40% longer than usual");
         }
 
         private static void CreateDummyScriptWithImporter(string path)
@@ -72,17 +75,41 @@ namespace Tests
         [TestInitialize]
         public void Initialize()
         {
+            CleanupScriptsDirectory();
+            CleanupMetaDataDirectory();
+
             MouseRobot.Instance.SetupProjectPath(TempProjectPath);
+            AssetManager.Instance.Refresh();
         }
 
         [TestCleanup]
         public void Cleanup()
         {
-            DirectoryInfo di = new DirectoryInfo(TempProjectPath + "\\" + AssetManager.ScriptFolder);
-            foreach (FileInfo file in di.GetFiles())
-                file.Delete();
+            CleanupScriptsDirectory();
+            CleanupMetaDataDirectory();
 
+            AssetGuidManager.Instance.LoadMetaFiles();
             AssetManager.Instance.Refresh();
+        }
+
+        private void CleanupMetaDataDirectory()
+        {
+            if (Directory.Exists(TempProjectPath + "\\" + AssetGuidManager.MetadataFolder))
+            {
+                DirectoryInfo di = new DirectoryInfo(TempProjectPath + "\\" + AssetGuidManager.MetadataFolder);
+                foreach (FileInfo file in di.GetFiles())
+                    file.Delete();
+            }
+        }
+
+        private void CleanupScriptsDirectory()
+        {
+            if (Directory.Exists(TempProjectPath + "\\" + AssetManager.ScriptFolder))
+            {
+                DirectoryInfo di = new DirectoryInfo(TempProjectPath + "\\" + AssetManager.ScriptFolder);
+                foreach (FileInfo file in di.GetFiles())
+                    file.Delete();
+            }
         }
     }
 }
